requirements.txt
asgiref==3.5.2
Django==4.1
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
windows-curses==2.3.0
